Fix many thread and file descriptor issues on Windows
The greenbean web server now works nearly perfectly on Windows with over
1000 threads. But some synchronization issues still remain which prevent
us from going over nine thousand.

+++ libc/calls/internal.h
@@ -66,7 +66,7 @@ struct Fd {
 };
 
 struct Fds {
-  size_t f; /* lowest free slot */
+  int f;    /* lowest free slot */
   size_t n; /* monotonic capacity */
   struct Fd *p;
   struct Fd __init_p[OPEN_MAX];
